mocked provider moved to providers/__mocks__
