+++ client/src/providers/__tests__/MockedProvider.test.js
@@ -2,21 +2,8 @@
 import React from 'react';
 import { render, screen, cleanup } from '@testing-library/react';
 import { test, expect, describe, afterEach, beforeEach } from '@jest/globals';
+import MockedProvider from 'providers/__mocks__/MockedProvider'
 import '@testing-library/jest-dom';
-
-type MockProviderT = {|
-  context: React$Context<any>,
-  children: any,
-  value: any,
-|};
-
-const MockedProvider = (props: MockProviderT): React$Element<any> => {
-  return (
-    <props.context.Provider value={props.value}>
-      {props.children}
-    </props.context.Provider>
-  );
-}
 
 describe('Mocked provider', () => {
   test('Mocked provider renders wihtout crashing', () => {
@@ -29,6 +16,3 @@ describe('Mocked provider', () => {
     expect(screen.getByText(/Test/i)).toBeInTheDocument();
   });
 });
-
-
-export default MockedProvider;
